A bunch of enhancements:
- build-victor.sh handles dependecy-getting. build-v.sh does much less stuff now
- build-victor.sh gets Go, UPX, protoc-gen-go stuff, vicos-sdk, and CMake
- Fix a couple Python2 formatting errors
- Docker container now works
- Check if protocCppPlugin works. If not, recompile it
- Branch now automatically makes its way into CCIS, no need to manually change it
- Docker might not be required anymore. I was able to build it, without error, on my Arch machine directly
- Still need to see if it all works on macOS
- Generally clean up logging
- The Makefiles generator now works if you wanna try that for some reason. Cool colors
- Use modcacherw Go flag so that a normal user can delete the gocache folder
- Restructure cache so it all lives under build/cache, mainly for a smooth transition to modcacherw-ed Go
- Always assume Release build, not Debug
- Update README

+++ build/build-v.sh
@@ -4,41 +4,53 @@ CURRENT_BUILDER="vic-standalone-builder-7"
 
 set -e
 
+if [[ $(id -u) == 0 ]]; then
+	echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
+    echo -e "\033[1;31mDo not run this script as root.\033[0m"
+    echo "If Docker is giving you a permission denied error, look over the README one more time. It includes instructions to allow Docker to be run as a normal user."
+	echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!"
+    exit 1
+fi
+
 if [[ "$(uname -a)" == *"Darwin"* ]]; then
-	echo "Installing required Go stuff..."
-	go install google.golang.org/protobuf/cmd/protoc-gen-go@latest
-	go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@latest
-	go install github.com/grpc-ecosystem/grpc-gateway/v2/protoc-gen-grpc-gateway@latest
-	export PATH="$PATH:$HOME/go/bin"
-	./project/victor/scripts/victor_build_release.sh "$@"
-	echo
-	echo -e "\033[1;32mComplete.\033[0m"
-	echo
+    ./project/victor/scripts/victor_build_release.sh "$@"
+    echo
+    echo -e "\033[1;32mComplete.\033[0m"
+    echo
 else
-	mkdir -p build/cache
-	mkdir -p build/gocache
-	mkdir -p build/usercache
-	if [[ -z $(docker images -q $CURRENT_BUILDER) ]]; then
-		docker build \
-		--build-arg DIR_PATH="$(pwd)" \
-		--build-arg USER_NAME=$USER \
-		--build-arg UID=$(id -u $USER) \
-		--build-arg GID=$(id -u $USER) \
-		-t $CURRENT_BUILDER \
-		build/
-	else
-		echo "Reusing $CURRENT_BUILDER"
-	fi
-	docker run --rm -it \
-		-v $(pwd)/anki-deps:/home/$USER/.anki \
-		-v $(pwd):$(pwd) \
-		-v $(pwd)/build/cache:/home/$USER/.ccache \
-		-v $(pwd)/build/gocache:/home/$USER/go \
-		-v $(pwd)/build/usercache:/home/$USER/.cache \
-		$CURRENT_BUILDER bash -c \
-		"cd $(pwd) && \
+    if [[ -d build/cache/0 ]]; then
+        echo "Rebuilding cache..."
+        rm -rf build/cache
+        # permissions prevent us from deleting as sudo. this
+        # is fixed in build/cache/go thanks to -modcacherw
+        #rm -rf build/gocache
+        rm -rf build/usercache
+    fi
+    mkdir -p anki-deps
+    mkdir -p build/cache/ccache
+    mkdir -p build/cache/go
+    mkdir -p build/cache/user
+    if [[ -z $(docker images -q $CURRENT_BUILDER) ]]; then
+        docker build \
+        --build-arg DIR_PATH="$(pwd)" \
+        --build-arg USER_NAME=$USER \
+        --build-arg UID=$(id -u $USER) \
+        --build-arg GID=$(id -u $USER) \
+        -t $CURRENT_BUILDER \
+        build/
+    else
+        echo "Reusing $CURRENT_BUILDER"
+    fi
+    docker run --rm -it \
+    -v $(pwd)/anki-deps:/home/$USER/.anki \
+    -v $(pwd):$(pwd) \
+    -v $(pwd)/build/cache/ccache:/home/$USER/.ccache \
+    -v $(pwd)/build/cache/go:/home/$USER/go \
+    -v $(pwd)/build/cache/user:/home/$USER/.cache \
+    $CURRENT_BUILDER bash -c \
+    "cd $(pwd) && \
 		./project/victor/scripts/victor_build_release.sh $@ && \
 		echo && \
 		echo -e \"\e[1;32mComplete.\e[0m\" && \
-		echo"
+    echo"
 fi
